requirements-barlow.txt
alabaster==0.7.16
albumentations==1.3.1
annotated-types==0.6.0
anyio==4.2.0
apeye==1.4.1
apeye-core==1.1.5
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autodocsumm==0.2.12
Babel==2.14.0
beautifulsoup4==4.12.3
black==24.1.1
bleach==6.1.0
bokeh==3.3.4
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
comm==0.2.1
contourpy==1.2.0
coverage==7.4.1
crashtest==0.4.1
cryptography==42.0.2
cssutils==2.9.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dicomweb-client==0.59.1
dict2css==0.3.0.post1
distlib==0.3.8
docker-pycreds==0.4.0
docutils==0.20.1
domdf-python-tools==3.8.0.post2
dulwich==0.21.7
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
fitter==1.7.0
Flask==2.2.5
Flask-Cors==4.0.0
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
furo==2024.1.29
gitdb==4.0.11
GitPython==3.1.41
Glymur==0.12.9.post1
greenlet==3.0.3
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.20.3
identify==2.5.33
idna==3.6
imagecodecs==2024.1.1
imageio==2.33.1
imagesize==1.4.1
importlib-metadata==7.0.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.0
ipython==8.21.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.0.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-cache==1.0.0
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
mypy==1.8.0
mypy-extensions==1.0.0
myst-nb==1.0.0
myst-parser==2.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.15
nodeenv==1.8.0
notebook_shim==0.2.3
numba==0.59.0
numcodecs==0.12.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openslide-python==1.3.1
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.4.0
poetry==1.7.1
poetry-bumpversion==0.3.2
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pre-commit==3.6.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydicom==2.4.4
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-runner==6.0.1
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
rapidfuzz==3.6.1
readme-renderer==42.0
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rich-click==1.8.2
rpds-py==0.17.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
ruff==0.1.13
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.2
sentry-sdk==1.40.0
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-autodoc-typehints==1.25.3
sphinx-basic-ng==1.0.0b2
sphinx-copybutton==0.5.2
sphinx-jinja2-compat==0.2.0.post1
sphinx-prompt==1.8.0
sphinx-tabs==3.4.5
sphinx-toolbox==3.5.0
sphinx_design==0.5.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==0.9.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.2.0
trove-classifiers==2024.1.31
twine==4.0.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
umap-learn==0.5.5
uri-template==1.3.0
urllib3==2.2.0
virtualenv==20.25.0
wandb==0.16.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsidicom==0.17.0
xyzservices==2023.10.1
zarr==2.16.1
zipp==3.17.0
